backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
psycopg2-binary==2.9.9
bleach==6.1.0